refactor(#3): 디렉토리 구조 변경
- atoms만 구성하는 것으로 하고 다른 컴포넌트는 관련 기능으로 묶는 것으로 변경
