+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ToDoListContainerProvider } from './contexts/ToDoListContainer';
-import ToDoListCard from './components/ToDoListCard';
+import ToDoListCard from './components/ToDo/ToDoListCard';
 
 function App(): JSX.Element {
   return (
